backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
pypdf2==3.0.1
python-pptx==0.6.23
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
redis==5.0.1
celery==5.3.4